+++ src/index.ts
@@ -1,4 +1,5 @@
 import { CuiInit } from './app/initializer';
-//import '../styles/style.scss';
+//import "../node_modules/cui-styles/styles/style.scss";
+import './styles.scss';
 const $: any = window;
 $.cuiInit = new CuiInit();